Get rid of the ItemUriFixture and tidy up the other items generators
